refactor(types): Diagnosis 类型集上移到 src/types/ 解开 observability 反向依赖
将 Diagnosis / DiagnosisBundle 等纯类型定义从 src/diagnosis/types.ts 上移到
src/types/diagnosis.ts(types/ barrel 同步 export),src/diagnosis/types.ts 缩为
类型 re-export + isActiveDiagnosisLifecycle 运行时 helper。observability/inbox.ts
对 DiagnosisBundle 的 type-only import 改走 ../types/diagnosis.js,observability →
diagnosis 反向依赖彻底切除。其余 server / diagnosis 内部消费者经 re-export 透明
保持,无外部 API 变化。
